Move store code to MockTask Move the Provider and mock store into MockTask, along with code that initialises and updates the store.
