+++ packages/lib-classifier/src/components/Classifier/components/TaskArea/components/Tasks/Tasks.stories.js
@@ -2,16 +2,7 @@ import { withKnobs, boolean, radios, select } from '@storybook/addon-knobs'
 import asyncStates from '@zooniverse/async-states'
 import { storiesOf } from '@storybook/react'
 import React from 'react'
-import { Provider } from 'mobx-react'
 import { MockTask } from '@helpers'
-import { createStore }  from '@store/helpers'
-
-const store = createStore()
-function addStepToStore (step, tasks) {
-  const steps = [ ['S1', step] ]
-  store.workflowSteps.setStepsAndTasks({ steps, tasks })
-  store.workflowSteps.active.tasks.forEach(task => store.classifications.addAnnotation(task))
-}
 
 storiesOf('Tasks / General', module)
   .addDecorator(withKnobs)
@@ -22,20 +13,16 @@ storiesOf('Tasks / General', module)
   })
   .add('loading', function () {
     return (
-      <Provider classifierStore={{}}>
-        <MockTask
-          loadingState={asyncStates.loading}
-        />
-      </Provider>
+      <MockTask
+        loadingState={asyncStates.loading}
+      />
     )
   })
   .add('error', function () {
     return (
-      <Provider classifierStore={{}}>
-        <MockTask
-          loadingState={asyncStates.error}
-        />
-      </Provider>
+      <MockTask
+        loadingState={asyncStates.error}
+      />
     )
   })
   .add('multiple tasks', function () {
@@ -61,21 +48,16 @@ storiesOf('Tasks / General', module)
       stepKey: 'S1',
       taskKeys: ['init', 'T1']
     }
-    addStepToStore(step, tasks)
     const dark = boolean('Dark theme', false)
     const subjectReadyState = select('Subject loading', asyncStates, asyncStates.success)
     const isThereTaskHelp = boolean('Enable task help', true)
     return (
-      <Provider classifierStore={store}>
-        <MockTask
-          dark={dark}
-          classification={store.classifications.active}
-          isThereTaskHelp={isThereTaskHelp}
-          loadingState={asyncStates.success}
-          step={store.workflowSteps.active}
-          subjectReadyState={subjectReadyState}
-          subjectViewer={store.subjectViewer}
-        />
-      </Provider>
+      <MockTask
+        dark={dark}
+        isThereTaskHelp={isThereTaskHelp}
+        step={step}
+        subjectReadyState={subjectReadyState}
+        tasks={tasks}
+      />
     )
   })
