+++ packages/lib-classifier/src/helpers/MockTask/MockTask.js
@@ -1,14 +1,26 @@
 import { Box, Grommet } from 'grommet'
-import { observer } from 'mobx-react'
+import { Provider, observer } from 'mobx-react'
 import React from 'react'
 import { Tasks } from '../../components/Classifier/components/TaskArea/components/Tasks/Tasks'
 import asyncStates from '@zooniverse/async-states'
 import zooTheme from '@zooniverse/grommet-theme'
+import { createStore }  from '@store/helpers'
+
+let store
 
 const ObservedTasks = observer(Tasks)
 
+function addStepToStore(step, tasks) {
+  const steps = [['S1', step]]
+  store.workflowSteps.setStepsAndTasks({ steps, tasks })
+  store.workflowSteps.active.tasks.forEach(task => store.classifications.addAnnotation(task))
+}
+
 export default function MockTask(props) {
-  const { dark, subjectViewer, ...taskProps } = props
+  const { dark, step, tasks, ...taskProps } = props
+  store = store ?? createStore()
+  addStepToStore(step, tasks)
+  const { subjectViewer } = store
   switch (taskProps.subjectReadyState) {
     case asyncStates.error: {
       subjectViewer.onError()
@@ -25,26 +37,32 @@ export default function MockTask(props) {
   }
 
   return (
-    <Grommet
-      background={{
-        dark: 'dark-1',
-        light: 'light-1'
-      }}
-      theme={Object.assign({}, zooTheme, { dark })}
-      themeMode={(dark) ? 'dark' : 'light'}
-    >
-      <Box
+    <Provider classifierStore={store}>
+      <Grommet
         background={{
-          dark: 'dark-3',
-          light: 'neutral-6'
+          dark: 'dark-1',
+          light: 'light-1'
         }}
-        pad='1em'
-        width='380px'
+        theme={Object.assign({}, zooTheme, { dark })}
+        themeMode={(dark) ? 'dark' : 'light'}
       >
-        <ObservedTasks
-          {...taskProps}
-        />
-      </Box>
-    </Grommet>
+        <Box
+          background={{
+            dark: 'dark-3',
+            light: 'neutral-6'
+          }}
+          pad='1em'
+          width='380px'
+        >
+          <ObservedTasks
+            classification={store.classifications.active}
+            loadingState={asyncStates.success}
+            step={store.workflowSteps.active}
+            subjectViewer={store.subjectViewer}
+            {...taskProps}
+          />
+        </Box>
+      </Grommet>
+    </Provider>
   )
 }
